And again, tidying things up in classes, as well as adhering to some naming conventions, etc.

+++ mos/host_conf.py
@@ -1,49 +1,74 @@
 #!/usr/bin/python3
 
-from mos import *
+import mos.helper as helper
+
 import subprocess
 import apt
 import shutil
 import os
 
-def host_tree_conf():
-	os.makedirs(mos_path + "/data/build", mode = 0o777, exist_ok = True)
-	os.makedirs(mos_path + "/data/build/kernel", mode = 0o777, exist_ok = True)
-	os.makedirs(mos_path + "/data/build/bootstrap", mode = 0o777, exist_ok = True)
-	os.makedirs(mos_path + "/data/images", mode = 0o777, exist_ok = True)
-	os.makedirs(mos_path + "/data/ssh_keys", mode = 0o777, exist_ok = True)
+class HostConf:
+	def __init__(self):
+		h = helper.Helper()
+		self.mos_path = h.mos_path
+		print(self.mos_path)
 
 
-def host_apt_conf():
-	systemAdmin = ["iproute2", "nftables"]
-	systemQemu = ["qemu-system", "libvirt-clients", "libvirt-daemon-system"]
-	system_build = ["build-ninja", "libpixman-1-dev", "meson"]
+	def tree_conf(self):
+		os.makedirs(self.mos_path + "/data/build", mode = 0o777, exist_ok = True)
+		os.makedirs(self.mos_path + "/data/build/kernel", mode = 0o777, exist_ok = True)
+		os.makedirs(self.mos_path + "/data/build/bootstrap", mode = 0o777, exist_ok = True)
+		os.makedirs(self.mos_path + "/data/images", mode = 0o777, exist_ok = True)
+		os.makedirs(self.mos_path + "/data/ssh_keys", mode = 0o777, exist_ok = True)
 
-	packageList = systemAdmin + systemQemu
 
-	cache = apt.cache.Cache()
-	cache.update()
-	cache.open()
+	def apt_conf(self):
+		systemAdmin = ["iproute2", "nftables"]
+		systemQemu = ["qemu-system", "libvirt-clients", "libvirt-daemon-	system"]
+		system_build = ["build-ninja", "libpixman-1-dev", "meson"]
 
-	for package in packageList:
+		packageList = systemAdmin + systemQemu
 
-		pkg_name = package
-		pkg = cache[pkg_name]
-		if pkg.is_installed:
-			print("{pkg_name} already installed".format(pkg_name=pkg_name))
+		cache = apt.cache.Cache()
+		cache.update()
+		cache.open()
+
+		for package in packageList:
+
+			pkg_name = package
+			pkg = cache[pkg_name]
+			if pkg.is_installed:
+				print("{pkg_name} already installed".format(pkg_name=pkg_name))
+			else:
+				pkg.mark_install()
+				cache.commit()
+
+
+	def grub_conf(self):
+		shutil.copyfile("conf/host/grub", "/etc/default/grub")
+		subprocess.run(["update-grub"])
+		print("grup-update was complete")
+
+
+	def syslink(self):
+		sys_link = "/usr/bin/meros"
+		
+		if os.path.exists(sys_link):
+			os.remove(sys_link)
 		else:
-			pkg.mark_install()
-			cache.commit()
+			pass
+		os.symlink(self.mos_path + "/meros.py", sys_link)
 
+	def main(self):
+		host_distro = "debian"
+		hc = HostConf()
 
-def host_grub_conf():
-	shutil.copyfile("conf/host/grub", "/etc/default/grub")
-	subprocess.run(["update-grub"])
-	print("grup-update was complete")
+		hc.tree_conf()
+		if host_distro == "ddebian":
+			HostConf.apt_conf(self)
+		else:
+			pass
 
-
-def host_syslink():
-	sys_link = "/usr/bin/meros"
-
-	os.remove(sys_link)
-	os.symlink(mos_path + "/meros.py", sys_link)
+		hc.grub_conf()
+		hc.syslink()
+
